fix bug for aa name and refactoring
