+++ shared-data/js/helpers/deckConfiguration/index.ts
@@ -1,0 +1,3 @@
+export * from './getFixtureFrom'
+export * from './getVisualSlotFrom'
+export * from './getVisualSlotFrom'
